+++ src/Engine.cpp
@@ -55,9 +55,6 @@ namespace Engine {
         // -----
         processInput(window);
 
-        //if (InputSystem::GetKeyDown(window, GLFW_KEY_ESCAPE))
-        //    std::cout << "escape1" << std::endl;
-
 
         //spdlog::info("frameEnd: {:03.20f}", frameEnd);
         //spdlog::info("frameStart: {:03.20f}", frameStart);