+++ src/InputSystem.cpp
@@ -5,13 +5,63 @@
 #include "../include/InputSystem.h"
 
 
-bool InputSystem::GetKeyDown(GLFWwindow* window, int key)
+
+void InputSystem::InputInit()
 {
-    if (glfwGetKey(window, key) == GLFW_PRESS)
+    window = Engine::getWindow();
+    glfwSetKeyCallback(window, key_callback);
+    glfwSetJoystickCallback(joystick_callback);
+}
+
+void InputSystem::update() {
+    lastKeyStates = keyStates;
+    keyStates.clear();
+
+    for (const auto& key : monitoredKeys) {
+        int state = glfwGetKey(window, key);
+        keyStates[key] = state == GLFW_PRESS || state == GLFW_REPEAT;
+    }
+}
+
+void InputSystem::monitorKey(int key) {
+    monitoredKeys.push_back(key);
+}
+
+bool InputSystem::GetKey(int key)
+{
+    return keyStates[key];
+}
+
+bool InputSystem::GetKeyDown(int key) {
+    return keyStates[key] && !lastKeyStates[key];
+}
+
+bool InputSystem::isGamepadButtonPressed(int button) {
+    int present = glfwJoystickPresent(GLFW_JOYSTICK_1);
+    if (present) {
+        int count;
+        const unsigned char* buttons = glfwGetJoystickButtons(GLFW_JOYSTICK_1, &count);
+        return buttons[button] == GLFW_PRESS;
+    }
+    return false;
+}
+
+float InputSystem::getJoystickAxis(int axis) {
+    int present = glfwJoystickPresent(GLFW_JOYSTICK_1);
+    if (present) {
+        int count;
+        const float* axes = glfwGetJoystickAxes(GLFW_JOYSTICK_1, &count);
+        return axes[axis];
+    }
+    return 0.0f;
+}
+
+float InputSystem::GetAxis(int key)
+{
+    GLFWgamepadstate state;
+    if (glfwGetGamepadState(GLFW_JOYSTICK_3, &state))
     {
-        return true;
+        return state.axes[key];
     }
-    else{
-        return false;
-    }
+    return 0;
 }