+++ changelog.md
@@ -21,7 +21,7 @@
 - MovingObject class
 - MovingObject has velocity in all axes, its position changes depending on deltaTime 
 - InputSystem class
-- InputSystem has a GetKeyDown method to determine whether key is pressed or not
+- InputSystem has a GetKey method to determine whether key is pressed or not
 
 ## [0.0.6] - 2023-03-31